refactor: . Move DateRange.ts and its tests to DateTime/ sub-directory
